# yandex/payment-server-gen/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Payment Service API
  description: API для проверки баланса и оформления заказа
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Локальный сервер

paths:
  /balance/{userId}:
    get:
      operationId: getBalanceById
      summary: Получение баланса
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
          description: Идентификатор пользователя
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: string
                    example: "1"
                  balance:
                    type: number
                    format: double
                    example: 1500.50
        '404':
          description: Пользователь не найден

  /payment:
    post:
      operationId: cartPayment
      summary: Оплата заказа
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userId
                - amount
              properties:
                userId:
                  type: string
                  example: "1"
                amount:
                  type: number
                  format: double
                  example: 500.00
      responses:
        '200':
          description: Платёж успешно выполнен
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "SUCCESS"
                  newBalance:
                    type: number
                    format: double
                    example: 1000.50
        '400':
          description: Недостаточно средств
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "FAILED"
        '404':
          description: Пользователь не найден