# yandex/payment-service/src/main/java/ru/yandex/payment/api/PaymentApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.5.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package ru.yandex.payment.api;

import ru.yandex.payment.model.CartPayment200Response;
import ru.yandex.payment.model.CartPayment400Response;
import ru.yandex.payment.model.CartPaymentRequest;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;
import org.springframework.web.server.ServerWebExchange;
import reactor.core.publisher.Flux;
import reactor.core.publisher.Mono;
import org.springframework.http.codec.multipart.Part;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-07-19T20:36:16.873307+04:00[Europe/Samara]", comments = "Generator version: 7.5.0")
@Validated
@Tag(name = "payment", description = "the payment API")
public interface PaymentApi {

    /**
     * POST /payment : Оплата заказа
     *
     * @param cartPaymentRequest  (required)
     * @return Платёж успешно выполнен (status code 200)
     *         or Недостаточно средств (status code 400)
     *         or Пользователь не найден (status code 404)
     */
    @Operation(
        operationId = "cartPayment",
        summary = "Оплата заказа",
        responses = {
            @ApiResponse(responseCode = "200", description = "Платёж успешно выполнен", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CartPayment200Response.class))
            }),
            @ApiResponse(responseCode = "400", description = "Недостаточно средств", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CartPayment400Response.class))
            }),
            @ApiResponse(responseCode = "404", description = "Пользователь не найден")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/payment",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default Mono<ResponseEntity<CartPayment200Response>> cartPayment(
        @Parameter(name = "CartPaymentRequest", description = "", required = true) @Valid @RequestBody Mono<CartPaymentRequest> cartPaymentRequest,
        @Parameter(hidden = true) final ServerWebExchange exchange
    ) {
        Mono<Void> result = Mono.empty();
        exchange.getResponse().setStatusCode(HttpStatus.NOT_IMPLEMENTED);
        for (MediaType mediaType : exchange.getRequest().getHeaders().getAccept()) {
            if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                String exampleString = "{ \"newBalance\" : 1000.5, \"status\" : \"SUCCESS\" }";
                result = ApiUtil.getExampleResponse(exchange, MediaType.valueOf("application/json"), exampleString);
                break;
            }
            if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                String exampleString = "{ \"status\" : \"FAILED\" }";
                result = ApiUtil.getExampleResponse(exchange, MediaType.valueOf("application/json"), exampleString);
                break;
            }
        }
        return result.then(cartPaymentRequest).then(Mono.empty());

    }

}